support Ovis1.6
